Added SSD model
As discussed in the chat the model is  not actually working properly but the pipeline should work. Please wait to run it because I need to check with a fresh mind if the function called from the util.py are behaving as expected. Comments on on how to structure better the pipeline are more than welcome.